backend/gateway/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
pydantic==2.8.2
python-jose[cryptography]==3.3.0
redis==5.0.7
